test_resources/test_server/django/requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.12.4
pytz==2021.3
sqlparse==0.4.2
gunicorn==20.1.0
